found the bug
In order_ok?, I was confusing parent and child indexes into @heap
with the *values* stored in the heap, which themselves (as PrimHeap
is used in this program) are are indices into a totally separate
array of edges.

This explains the strange behavior where removing any edge from
msf.in0, including an edge from a different component than the one
in which the representation invariant violation was detected,
prevented the bug from being shown.

I hope other serious bugs, if any, are easier for me to find.